+++ CHANGES.md
@@ -5,6 +5,8 @@ videojs-record changelog
 ------------------
 
 - Add support for keyboard hotkeys (requires video.js 7.5.0 or newer) (#339)
+- Trigger error when `getUserMedia` or `getDisplayMedia` support is missing (in
+  Chrome iOS for example) (#353)
 - Improve pause/resume accuracy with monotonic clock
 - Make sure recorded data is garbage collected
 - Bump required version for videojs-wavesurfer to 2.8.0 for wavesurfer.js fixes